Move cltv_multisig_data.h out of generated data directory.
